+++ src/components/community/Staked/components/UnstakingModal/hooks/useUnstakingModalMachine.ts
@@ -1,8 +1,12 @@
+import { parseEther } from "@ethersproject/units"
 import { useWeb3React } from "@web3-react/core"
 import { useMachine } from "@xstate/react"
+import { useCommunity } from "components/community/Context"
+import useBalance from "hooks/useBalance"
+import useContract from "hooks/useContract"
 import { useEffect } from "react"
 import { assign, createMachine, DoneInvokeEvent } from "xstate"
-import useUnstake from "./useUnstake"
+import AGORA_SPACE_ABI from "constants/agoraSpaceABI.json"
 
 type ContextType = {
   error: Error | null
@@ -52,9 +56,24 @@ const unstakingModalMachine = createMachine<ContextType, DoneInvokeEvent<any>>(
 )
 
 const useUnstakingModalMachine = (): any => {
-  const { unstake } = useUnstake()
+  const {
+    chainData: {
+      contract: { address },
+      stakeToken,
+    },
+  } = useCommunity()
+  const { data: staked } = useBalance(stakeToken)
+  const contract = useContract(address, AGORA_SPACE_ABI, true)
   const { account, chainId } = useWeb3React()
-  const [state, send] = useMachine(unstakingModalMachine, { services: { unstake } })
+  const [state, send] = useMachine(unstakingModalMachine, {
+    services: {
+      unstake: async () => {
+        const weiAmount = parseEther(staked.toString())
+        const tx = await contract.withdraw(weiAmount)
+        return tx
+      },
+    },
+  })
 
   useEffect(() => {
     send("RESET")
